Refactored period
